feat: move extract_google_doc_id to new utils module
