+++ app/integrations/google_next/docs.py
@@ -1,5 +1,7 @@
-import re
+"""Google Docs API integration."""
+
 from googleapiclient.discovery import Resource  # type: ignore
+
 from integrations.google_next.service import (
     execute_google_api_call,
     handle_google_api_errors,
@@ -10,6 +12,7 @@ from core.logging import get_module_logger
 DEFAULT_SCOPES = [
     "https://www.googleapis.com/auth/documents",
 ]
+
 logger = get_module_logger()
 
 
@@ -142,31 +145,3 @@ class GoogleDocs:
             documentId=document_id,
             **kwargs,
         )
-
-
-def extract_google_doc_id(url):
-    """
-    Extracts the Google Docs ID from a Google Docs URL.
-
-    Args:
-        url (str): The URL of the Google Docs document.
-
-    Returns:
-        str: The Google Docs ID extracted from the URL.
-    """
-    logger.debug(
-        "extracting_google_doc_id",
-        url=url,
-    )
-    if not url:
-        return None
-
-    # Regular expression pattern to match Google Docs ID
-    pattern = r"https://docs.google.com/document/d/([a-zA-Z0-9_-]+)/"
-
-    # Search in the given text for all occurences of pattern
-    match = re.search(pattern, url)
-    if match:
-        return match.group(1)
-    else:
-        return None
