+++ app/tests/integrations/google_next/test_docs.py
@@ -5,8 +5,7 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
-# from integrations.google_next.docs import docs.GoogleDocs
-from integrations.google_next import docs
+from integrations.google_next.docs import GoogleDocs
 
 
 @pytest.fixture
@@ -20,31 +19,23 @@ def mock_service():
 
 @pytest.fixture
 def google_docs(mock_service):
-    # Default usage: no arguments
-    return docs.GoogleDocs()
-
-
-@pytest.fixture
-def google_docs_with_email(mock_service):
-    scopes = ["https://www.googleapis.com/auth/documents"]
-    delegated_email = "email@test.com"
-    return docs.GoogleDocs(scopes, delegated_email)
-
-
-@pytest.fixture
-def google_docs_with_scopes(mock_service):
-    scopes = ["https://www.googleapis.com/auth/documents.readonly"]
-    return docs.GoogleDocs(scopes)
+    with patch("integrations.google_next.docs.DEFAULT_SCOPES", ["tests", "scopes"]):
+        return GoogleDocs()
 
 
 class TestGoogleDocs:
+    """Unit tests for the GoogleDocs class."""
+
     @pytest.fixture(autouse=True)
-    def setup(self, google_docs: docs.GoogleDocs):
+    # pylint: disable=redefined-outer-name
+    def setup(self, google_docs: GoogleDocs):
+        # pylint: disable=attribute-defined-outside-init
         self.google_docs = google_docs
 
+    # @patch("integrations.google_next.docs.DEFAULT_SCOPES", ["test", "scopes"])
     def test_init_uses_defaults(self):
-        """Test initialization without scopes and service raises ValueError."""
-        assert self.google_docs.scopes == docs.DEFAULT_SCOPES
+        """Test initialization with default scopes and no delegated email."""
+        assert self.google_docs.scopes == ["tests", "scopes"]
         assert self.google_docs.delegated_email is None
 
     @patch("integrations.google_next.docs.get_google_service")
@@ -53,7 +44,7 @@ class TestGoogleDocs:
     ):
         """Test initialization with delegated email and service."""
         mock_get_google_service.return_value = MagicMock()
-        google_docs = docs.GoogleDocs(
+        google_docs = GoogleDocs(
             scopes=["https://www.googleapis.com/auth/documents"],
             delegated_email="email@test.com",
             service=mock_get_google_service.return_value,
@@ -139,31 +130,3 @@ class TestGoogleDocs:
             documentId=document_id,
             something="else",
         )
-
-
-def test_extract_googe_doc_id_valid_google_docs_url():
-    url = "https://docs.google.com/document/d/1aBcD_efGHI/edit"
-    assert docs.extract_google_doc_id(url) == "1aBcD_efGHI"
-
-
-def test_extract_googe_doc_id_oogle_docs_url_with_parameters():
-    url = "https://docs.google.com/document/d/1aBcD_efGHI/edit?usp=sharing"
-    assert docs.extract_google_doc_id(url) == "1aBcD_efGHI"
-
-
-def test_extract_googe_doc_id_non_google_docs_url():
-    url = "https://www.example.com/page/d/1aBcD_efGHI/other"
-    assert docs.extract_google_doc_id(url) is None
-
-
-def test_extract_googe_doc_id_invalid_url_format():
-    url = "https://docs.google.com/document/1aBcD_efGHI"
-    assert docs.extract_google_doc_id(url) is None
-
-
-def test_extract_googe_doc_id_empty_string():
-    assert docs.extract_google_doc_id("") is None
-
-
-def test_extract_googe_doc_id_none_input():
-    assert docs.extract_google_doc_id(None) is None
